WIP, ENH: control reduction types
* still a work in progress--this branch can compile
and run `02_simple_reduce.py` with the `pk.int64` type,
but somewhere in the control flow an overflow is
still happening for large values compared to NumPy

* `python runtests.py` passes 87 tests and fails 2
on this feature branch:
```
FAILED tests/test_lambdas.py::TestLambda::test_add - SystemExit: 1
FAILED tests/test_lambdas.py::TestLambda::test_add_squares - SystemExit: 1
```
- they look like type-related failures due to my crude initial
work here

+++ examples/kokkos/02_simple_reduce.py
@@ -5,7 +5,7 @@ import pykokkos as pk
 class SquareSum:
     def __init__(self, n):
         self.N: int = n
-        self.total: pk.double = 0
+        self.total: pk.int64 = 0
 
     @pk.main
     def run(self):
@@ -16,7 +16,7 @@ class SquareSum:
         print("Sum:", self.total)
 
     @pk.workunit
-    def squaresum(self, i: int, acc: pk.Acc[pk.double]):
+    def squaresum(self, i: int, acc: pk.Acc[pk.int64]):
         acc += i * i
